moved common to conftest
