+++ services/api-server/tests/unit/api_solvers/conftest.py
@@ -5,13 +5,10 @@
 
 import json
 from copy import deepcopy
-from pathlib import Path
-from typing import Any, Iterator
+from typing import Any
 
 import httpx
 import pytest
-import respx
-import yaml
 from faker import Faker
 from fastapi import FastAPI, status
 from models_library.projects import Project
@@ -19,34 +16,6 @@ from models_library.utils.fastapi_encoders import jsonable_encoder
 from pytest_simcore.helpers import faker_catalog
 from respx import MockRouter
 from simcore_service_api_server.core.settings import ApplicationSettings
-
-
-@pytest.fixture
-def catalog_service_openapi_specs(osparc_simcore_services_dir: Path) -> dict[str, Any]:
-    openapi_path = osparc_simcore_services_dir / "catalog" / "openapi.json"
-    openapi_specs = json.loads(openapi_path.read_text())
-    return openapi_specs
-
-
-@pytest.fixture
-def directorv2_service_openapi_specs(
-    osparc_simcore_services_dir: Path,
-) -> dict[str, Any]:
-    return json.loads(
-        (osparc_simcore_services_dir / "director-v2" / "openapi.json").read_text()
-    )
-
-
-@pytest.fixture
-def webserver_service_openapi_specs(
-    osparc_simcore_services_dir: Path,
-) -> dict[str, Any]:
-    return yaml.safe_load(
-        (
-            osparc_simcore_services_dir
-            / "web/server/src/simcore_service_webserver/api/v0/openapi.yaml"
-        ).read_text()
-    )
 
 
 @pytest.fixture
@@ -122,54 +91,46 @@ def mocked_webserver_service_api(
 
 @pytest.fixture
 def mocked_catalog_service_api(
-    app: FastAPI, catalog_service_openapi_specs: dict[str, Any]
-) -> Iterator[MockRouter]:
-    settings: ApplicationSettings = app.state.settings
-    assert settings.API_SERVER_CATALOG
+    app: FastAPI,
+    mocked_catalog_service_api_base: MockRouter,
+    catalog_service_openapi_specs: dict[str, Any],
+) -> MockRouter:
 
+    respx_mock = mocked_catalog_service_api_base
     openapi = deepcopy(catalog_service_openapi_specs)
     schemas = openapi["components"]["schemas"]
 
-    # pylint: disable=not-context-manager
-    with respx.mock(
-        base_url=settings.API_SERVER_CATALOG.base_url,
-        assert_all_called=False,
-        assert_all_mocked=True,
-    ) as respx_mock:
-        respx_mock.get("/v0/meta").respond(200, json=schemas["Meta"]["example"])
+    respx_mock.get(
+        "/v0/services?user_id=1&details=false", name="list_services"
+    ).respond(
+        200,
+        json=[
+            # one solver
+            faker_catalog.create_service_out(
+                key="simcore/services/comp/Foo", name="Foo"
+            ),
+            # two version of the same solver
+            faker_catalog.create_service_out(version="0.0.1"),
+            faker_catalog.create_service_out(version="1.0.1"),
+            # not a solver
+            faker_catalog.create_service_out(type="dynamic"),
+        ],
+    )
 
-        # ----
-        respx_mock.get(
-            "/v0/services?user_id=1&details=false", name="list_services"
-        ).respond(
-            200,
-            json=[
-                # one solver
-                faker_catalog.create_service_out(
-                    key="simcore/services/comp/Foo", name="Foo"
-                ),
-                # two version of the same solver
-                faker_catalog.create_service_out(version="0.0.1"),
-                faker_catalog.create_service_out(version="1.0.1"),
-                # not a solver
-                faker_catalog.create_service_out(type="dynamic"),
-            ],
-        )
+    # -----
+    # NOTE: we could use https://python-jsonschema.readthedocs.io/en/stable/
+    #
 
-        # -----
-        # NOTE: we could use https://python-jsonschema.readthedocs.io/en/stable/
-        #
+    respx_mock.get(
+        # NOTE: regex does not work even if tested https://regex101.com/r/drVAGr/1
+        # path__regex=r"/v0/services/(?P<service_key>[\w/%]+)/(?P<service_version>[\d\.]+)/ports\?user_id=(?P<user_id>\d+)",
+        path__startswith="/v0/services/simcore%2Fservices%2Fcomp%2Fitis%2Fsleeper/2.1.4/ports",
+        name="list_service_ports",
+    ).respond(
+        200,
+        json=[
+            schemas["ServicePortGet"]["example"],
+        ],
+    )
 
-        respx_mock.get(
-            # NOTE: regex does not work even if tested https://regex101.com/r/drVAGr/1
-            # path__regex=r"/v0/services/(?P<service_key>[\w/%]+)/(?P<service_version>[\d\.]+)/ports\?user_id=(?P<user_id>\d+)",
-            path__startswith="/v0/services/simcore%2Fservices%2Fcomp%2Fitis%2Fsleeper/2.1.4/ports",
-            name="list_service_ports",
-        ).respond(
-            200,
-            json=[
-                schemas["ServicePortGet"]["example"],
-            ],
-        )
-
-        yield respx_mock
+    return respx_mock
